+++ packages/block-library/src/search/edit.js
@@ -453,7 +453,7 @@ export default function SearchEdit( {
 								return (
 									<Button
 										key={ widthValue }
-										isSmall
+										size="small"
 										variant={
 											widthValue === width &&
 											widthUnit === '%'