+++ packages/edit-post/src/store/test/actions.js
@@ -146,34 +146,6 @@ describe( 'actions', () => {
 		).toBe( true );
 	} );
 
-	describe( '__unstableSwitchToTemplateMode', () => {
-		it( 'welcome guide is active', () => {
-			// Activate `welcomeGuideTemplate` feature.
-			registry
-				.dispatch( editPostStore )
-				.toggleFeature( 'welcomeGuideTemplate' );
-			registry.dispatch( editPostStore ).__unstableSwitchToTemplateMode();
-			expect(
-				registry.select( editPostStore ).isEditingTemplate()
-			).toBeTruthy();
-			const notices = registry.select( noticesStore ).getNotices();
-			expect( notices ).toHaveLength( 0 );
-		} );
-
-		it( 'welcome guide is inactive', () => {
-			expect(
-				registry.select( editPostStore ).isEditingTemplate()
-			).toBeFalsy();
-			registry.dispatch( editPostStore ).__unstableSwitchToTemplateMode();
-			expect(
-				registry.select( editPostStore ).isEditingTemplate()
-			).toBeTruthy();
-			const notices = registry.select( noticesStore ).getNotices();
-			expect( notices ).toHaveLength( 1 );
-			expect( notices[ 0 ].content ).toMatch( 'template' );
-		} );
-	} );
-
 	describe( 'hideBlockTypes', () => {
 		it( 'adds the hidden block type to the preferences', () => {
 			registry
@@ -233,83 +205,6 @@ describe( 'actions', () => {
 		} );
 	} );
 
-	describe( 'toggleEditorPanelEnabled', () => {
-		it( 'toggles panels to be enabled and not enabled', () => {
-			// This will switch it off, since the default is on.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelEnabled( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelEnabled( 'control-panel' )
-			).toBe( false );
-
-			// Also check that the `getPreference` selector includes panels.
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {
-				'control-panel': {
-					enabled: false,
-				},
-			} );
-
-			// Switch it on again.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelEnabled( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelEnabled( 'control-panel' )
-			).toBe( true );
-
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {} );
-		} );
-	} );
-
-	describe( 'toggleEditorPanelOpened', () => {
-		it( 'toggles panels open and closed', () => {
-			// This will open it, since the default is closed.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelOpened( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelOpened( 'control-panel' )
-			).toBe( true );
-
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {
-				'control-panel': {
-					opened: true,
-				},
-			} );
-
-			// Close it.
-			registry
-				.dispatch( editPostStore )
-				.toggleEditorPanelOpened( 'control-panel' );
-
-			expect(
-				registry
-					.select( editPostStore )
-					.isEditorPanelOpened( 'control-panel' )
-			).toBe( false );
-
-			expect(
-				registry.select( editPostStore ).getPreference( 'panels' )
-			).toEqual( {} );
-		} );
-	} );
-
 	describe( 'updatePreferredStyleVariations', () => {
 		it( 'sets a preferred style variation for a block when a style name is passed', () => {
 			registry
@@ -365,7 +260,7 @@ describe( 'actions', () => {
 			registry
 				.dispatch( preferencesStore )
 				.set( 'core/edit-post', 'fixedToolbar', true );
-			registry.dispatch( editPostStore ).setIsListViewOpened( true );
+			registry.dispatch( editorStore ).setIsListViewOpened( true );
 			registry
 				.dispatch( editPostStore )
 				.openGeneralSidebar( 'edit-post/block' );
@@ -376,10 +271,10 @@ describe( 'actions', () => {
 					.select( preferencesStore )
 					.get( 'core/edit-post', 'fixedToolbar' )
 			).toBe( true );
-			expect( registry.select( editPostStore ).isListViewOpened() ).toBe(
+			expect( registry.select( editorStore ).isListViewOpened() ).toBe(
 				false
 			);
-			expect( registry.select( editPostStore ).isInserterOpened() ).toBe(
+			expect( registry.select( editorStore ).isInserterOpened() ).toBe(
 				false
 			);
 			expect(
@@ -392,20 +287,6 @@ describe( 'actions', () => {
 					.select( preferencesStore )
 					.get( 'core/edit-post', 'distractionFree' )
 			).toBe( true );
-		} );
-	} );
-
-	describe( 'setIsListViewOpened', () => {
-		it( 'should turn off distraction free mode when opening the list view', () => {
-			registry
-				.dispatch( preferencesStore )
-				.set( 'core/edit-post', 'distractionFree', true );
-			registry.dispatch( editPostStore ).setIsListViewOpened( true );
-			expect(
-				registry
-					.select( preferencesStore )
-					.get( 'core/edit-post', 'distractionFree' )
-			).toBe( false );
 		} );
 	} );
 } );