+++ packages/edit-site/src/components/preferences-modal/index.js
@@ -11,6 +11,7 @@ import { useMemo } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { useSelect, useDispatch, useRegistry } from '@wordpress/data';
 import { store as preferencesStore } from '@wordpress/preferences';
+import { store as editorStore } from '@wordpress/editor';
 
 /**
  * Internal dependencies
@@ -28,8 +29,9 @@ export default function EditSitePreferencesModal() {
 	const toggleModal = () =>
 		isModalActive ? closeModal() : openModal( PREFERENCES_MODAL_NAME );
 	const registry = useRegistry();
-	const { closeGeneralSidebar, setIsListViewOpened, setIsInserterOpened } =
-		useDispatch( editSiteStore );
+	const { closeGeneralSidebar } = useDispatch( editSiteStore );
+	const { setIsListViewOpened, setIsInserterOpened } =
+		useDispatch( editorStore );
 
 	const { set: setPreference } = useDispatch( preferencesStore );
 	const toggleDistractionFree = () => {
@@ -41,17 +43,58 @@ export default function EditSitePreferencesModal() {
 		} );
 	};
 
+	const turnOffDistractionFree = () => {
+		setPreference( 'core/edit-site', 'distractionFree', false );
+	};
+
 	const sections = useMemo( () => [
 		{
 			name: 'general',
 			tabLabel: __( 'General' ),
 			content: (
+				<PreferencesModalSection title={ __( 'Interface' ) }>
+					<EnableFeature
+						featureName="showListViewByDefault"
+						help={ __(
+							'Opens the block list view sidebar by default.'
+						) }
+						label={ __( 'Always open list view' ) }
+					/>
+					<EnableFeature
+						featureName="showBlockBreadcrumbs"
+						help={ __(
+							'Shows block breadcrumbs at the bottom of the editor.'
+						) }
+						label={ __( 'Display block breadcrumbs' ) }
+					/>
+					<EnableFeature
+						featureName="allowRightClickOverrides"
+						help={ __(
+							'Allows contextual list view menus via right-click, overriding browser defaults.'
+						) }
+						label={ __( 'Allow right-click contextual menus' ) }
+					/>
+				</PreferencesModalSection>
+			),
+		},
+		{
+			name: 'appearance',
+			tabLabel: __( 'Appearance' ),
+			content: (
 				<PreferencesModalSection
 					title={ __( 'Appearance' ) }
 					description={ __(
-						'Customize options related to the block editor interface and editing flow.'
+						'Customize the editor interface to suit your needs.'
 					) }
 				>
+					<EnableFeature
+						featureName="fixedToolbar"
+						onToggle={ turnOffDistractionFree }
+						help={ __(
+							'Access all block and document tools in a single place.'
+						) }
+						label={ __( 'Top toolbar' ) }
+					/>
 					<EnableFeature
 						featureName="distractionFree"
 						onToggle={ toggleDistractionFree }
@@ -67,46 +110,38 @@ export default function EditSitePreferencesModal() {
 						) }
 						label={ __( 'Spotlight mode' ) }
 					/>
-					<EnableFeature
-						featureName="showIconLabels"
-						label={ __( 'Show button text labels' ) }
-						help={ __( 'Show text instead of icons on buttons.' ) }
-					/>
-					<EnableFeature
-						featureName="showListViewByDefault"
-						help={ __(
-							'Opens the block list view sidebar by default.'
-						) }
-						label={ __( 'Always open list view' ) }
-					/>
-					<EnableFeature
-						featureName="showBlockBreadcrumbs"
-						help={ __(
-							'Shows block breadcrumbs at the bottom of the editor.'
-						) }
-						label={ __( 'Display block breadcrumbs' ) }
-					/>
 				</PreferencesModalSection>
 			),
 		},
 		{
-			name: 'blocks',
-			tabLabel: __( 'Blocks' ),
+			name: 'accessibility',
+			tabLabel: __( 'Accessibility' ),
 			content: (
-				<PreferencesModalSection
-					title={ __( 'Block interactions' ) }
-					description={ __(
-						'Customize how you interact with blocks in the block library and editing canvas.'
-					) }
-				>
-					<EnableFeature
-						featureName="keepCaretInsideBlock"
-						help={ __(
-							'Aids screen readers by stopping text caret from leaving blocks.'
+				<>
+					<PreferencesModalSection
+						title={ __( 'Navigation' ) }
+						description={ __(
+							'Optimize the editing experience for enhanced control.'
 						) }
-						label={ __( 'Contain text cursor inside block' ) }
-					/>
-				</PreferencesModalSection>
+					>
+						<EnableFeature
+							featureName="keepCaretInsideBlock"
+							help={ __(
+								'Keeps the text cursor within the block boundaries, aiding users with screen readers by preventing unintentional cursor movement outside the block.'
+							) }
+							label={ __( 'Contain text cursor inside block' ) }
+						/>
+					</PreferencesModalSection>
+					<PreferencesModalSection title={ __( 'Interface' ) }>
+						<EnableFeature
+							featureName="showIconLabels"
+							label={ __( 'Show button text labels' ) }
+							help={ __(
+								'Show text instead of icons on buttons across the interface.'
+							) }
+						/>
+					</PreferencesModalSection>
+				</>
 			),
 		},
 	] );