+++ packages/edit-widgets/src/components/header/document-tools/index.js
@@ -22,9 +22,9 @@ import useLastSelectedWidgetArea from '../../../hooks/use-last-selected-widget-a
 import { store as editWidgetsStore } from '../../../store';
 import { unlock } from '../../../lock-unlock';
 
-const { useShouldContextualToolbarShow } = unlock( blockEditorPrivateApis );
+const { useCanBlockToolbarBeFocused } = unlock( blockEditorPrivateApis );
 
-function DocumentTools( { setListViewToggleElement } ) {
+function DocumentTools() {
 	const isMediumViewport = useViewportMatch( 'medium' );
 	const inserterButton = useRef();
 	const widgetAreaClientId = useLastSelectedWidgetArea();
@@ -35,14 +35,18 @@ function DocumentTools( { setListViewToggleElement } ) {
 			),
 		[ widgetAreaClientId ]
 	);
-	const { isInserterOpen, isListViewOpen } = useSelect( ( select ) => {
-		const { isInserterOpened, isListViewOpened } =
-			select( editWidgetsStore );
-		return {
-			isInserterOpen: isInserterOpened(),
-			isListViewOpen: isListViewOpened(),
-		};
-	}, [] );
+	const { isInserterOpen, isListViewOpen, listViewToggleRef } = useSelect(
+		( select ) => {
+			const { isInserterOpened, isListViewOpened, getListViewToggleRef } =
+				unlock( select( editWidgetsStore ) );
+			return {
+				isInserterOpen: isInserterOpened(),
+				isListViewOpen: isListViewOpened(),
+				listViewToggleRef: getListViewToggleRef(),
+			};
+		},
+		[]
+	);
 	const { setIsWidgetAreaOpen, setIsInserterOpened, setIsListViewOpened } =
 		useDispatch( editWidgetsStore );
 	const { selectBlock } = useDispatch( blockEditorStore );
@@ -71,17 +75,8 @@ function DocumentTools( { setListViewToggleElement } ) {
 		[ setIsListViewOpened, isListViewOpen ]
 	);
 
-	const {
-		shouldShowContextualToolbar,
-		canFocusHiddenToolbar,
-		fixedToolbarCanBeFocused,
-	} = useShouldContextualToolbarShow();
 	// If there's a block toolbar to be focused, disable the focus shortcut for the document toolbar.
-	// There's a fixed block toolbar when the fixed toolbar option is enabled or when the browser width is less than the large viewport.
-	const blockToolbarCanBeFocused =
-		shouldShowContextualToolbar ||
-		canFocusHiddenToolbar ||
-		fixedToolbarCanBeFocused;
+	const blockToolbarCanBeFocused = useCanBlockToolbarBeFocused();
 
 	return (
 		<NavigableToolbar
@@ -119,7 +114,7 @@ function DocumentTools( { setListViewToggleElement } ) {
 						/* translators: button label text should, if possible, be under 16 characters. */
 						label={ __( 'List View' ) }
 						onClick={ toggleListView }
-						ref={ setListViewToggleElement }
+						ref={ listViewToggleRef }
 					/>
 				</>
 			) }